+++ stream.js
@@ -1,25 +1,26 @@
 // TODO Individually hide video/char per column
-// TODO Embed new twitch js library to make managing video/chat easier?
 
 class Stream {
     /**
      * Create a stream
-     * @param {string} username - Username of the channel.
-     * @param {boolean} doAPICalls - periodically call APIs for viewer count, live status, etc.
+     * @param {Object} streamManager StreamManager class this will be apart of
+     * @param {string} username Username of the channel
+     * @param {boolean} doAPICalls periodically call APIs for viewer count, live status, etc.
      */
-    constructor(username, doAPICalls = true) {
+    constructor(streamManager, username, doAPICalls = true) {
         if(username == "") { throw "Username empty"; }
 
+        this._streamManager = streamManager;
         this._username = username.trim();
         this._platform = "twitch";
 
-        // Initially null to prevent unneccesary background from iframe JS
+        // Initially null to prevent unneccesary background iframe JS
+        this._column = null;
         this._player = null;
         this._banner = null;
         this._chat = null;
         this._resizeChat = false;
         this._doAPICalls = doAPICalls;
-        this._apiTimeout = null;  // Stores setTimeout reference for API calls
     }
 
     /**
@@ -32,29 +33,7 @@ class Stream {
      * Key will expire in a month or two and will need to be replaced
      * Get new one with POST https://id.twitch.tv/oauth2/token?grant_type=client_credentials&client_id={}&client_secret={}
      */
-    static get TWITCH_OAUTH_ID() { return "jotxtukprcgp4zwmosrd2tlpullwx8"; }
-
-    /**
-     * Create a hh:ss display (or dd:hh:ss if >24 hours) for a stream's uptime
-     * @param {*} startTime the time a stream started. Must be parsable by JS's Date object
-     */
-    static get_uptime(startTime) {
-        let sec = Math.abs(new Date(startTime).getTime() - Date.now());
-        let mins = Math.floor(sec / 60000);
-        let hrs = Math.floor(mins / 60);
-        let days = Math.floor(hrs / 24);
-        mins = mins % 60;
-        hrs = hrs % 24;
-
-        days = days.toString().padStart(2, "0");
-        hrs = hrs.toString().padStart(2, "0");
-        mins = mins.toString().padStart(2, "0");
-        if (days == 0) {
-            return `${hrs}:${mins}`;
-        } else {
-            return `${days}:${hrs}:${mins}`;
-        }
-    }
+    static get TWITCH_OAUTH_ID() { return "37ulc0tjj2hvz9hvjq2qpsf60mdzta"; }
 
     /**
      * Return username of streamer
@@ -107,9 +86,9 @@ class Stream {
      * Ger url to a channel
      */
     getChannelURL() {
-        switch (this._platform) {
+        switch (this.getPlatform()) {
             case "twitch":
-                return "https://twitch.tv/" + this._username;
+                return "https://twitch.tv/" + this.getUsername();
         }
     }
 
@@ -119,7 +98,7 @@ class Stream {
     getVideoURL() {
         switch (this._platform) {
             case "twitch":
-                return `https://player.twitch.tv/?channel=${this._username}&parent=${document.location.hostname}`;
+                return `https://player.twitch.tv/?channel=${this.getUsername()}&parent=${document.location.hostname}`;
         }
     }
 
@@ -129,7 +108,7 @@ class Stream {
     getChatURL() {
         switch (this._platform) {
             case "twitch":
-                return `https://www.twitch.tv/embed/${this._username}/chat?parent=${document.location.hostname}&darkpopout`;
+                return `https://www.twitch.tv/embed/${this.getUsername()}/chat?parent=${document.location.hostname}&darkpopout`;
         }
     }
 
@@ -149,75 +128,100 @@ class Stream {
         return iframe;
     }
 
+    getColumn(settings) {
+        if(this._column) {
+            return this._column;
+        }
+
+        let div = htmlToElement("<div class='col'></div>");
+        if (settings.indexOf("novideo") == -1) { div.appendChild(this.getPlayer()); }
+        if (settings.indexOf("nobanner") == -1) { div.appendChild(this.getBanner()); }
+        if (settings.indexOf("nochat") == -1) { div.appendChild(this.getChat()); }
+
+        this._column = div;
+        return div;
+    }
+
     /**
      * Generates a streamer "banner" which contains link to
      * streamer page along with dom elements for _runAPICalls()
      */
     _genBanner() {
         // TODO Option to hide video/chat per column
-        let banner = htmlToElement(`<div class='banner' style='background: var(--${this._platform}-color);'></div>`);
+        let banner = htmlToElement(`<div class='banner' style='background: var(--${this.getPlatform()}-color);'></div>`);
         let bannerTop = htmlToElement("<div class='top'></div>");
         let bannerBot = htmlToElement("<div class='bot hidden'></div>");
         let channelName = htmlToElement(`<span class='channelName'>${this.getUsername()}</span>`);
         let updateCols = htmlToElement("<div class='rightWrapper'></div>");
         let addStreamIcon = htmlToElement("<a title='Add another stream' class='addStream icon'>&#x2795</a>");
         let removeStreamIcon = htmlToElement("<a title='Remove this stream' class='removeStream icon'>&#x2796;</a>");
-        removeStreamIcon.onclick = () => removeDOMStream(this.getUsername(), window.streamColumns);
+        removeStreamIcon.addEventListener("click", () =>  this._streamManager.removeStream(this.getUsername()));
 
-        let streamInput = htmlToElement(`<div class="inner-input-group rightWrapper">
-                <input type="text" placeholder="Username" required>
-            </div>`);
+        let streamInputWrapper = htmlToElement(`<div class="inner-input-group rightWrapper"></div>`);
+        let streamInput = htmlToElement(`<input type="text" placeholder="Username" required>`);
+        streamInput.addEventListener("keydown", (e) => {
+            let inputStream = streamInput.value.trim();
+            switch(e.key) {
+                case "Enter": {
+                    this._streamManager.addStream(inputStream, this.getUsername());
+
+                    bannerBot.classList.add("hidden");
+                    streamInput.value = "";
+                    break;
+                }
+
+                case "Escape": {
+                    bannerBot.classList.add("hidden");
+                    streamInput.value = "";
+                    break;
+                }
+            }
+        });
+
         let viewerCount = htmlToElement("<span class='viewerCount'></span>");
         let isLiveIcon = htmlToElement("<div class='liveIcon'></div>");
 
         let streamInputCancel = htmlToElement("<button>Cancel</button>");
-        streamInputCancel.onclick = () => bannerBot.classList.add("hidden");
+        streamInputCancel.addEventListener("click", () => bannerBot.classList.add("hidden"));
 
         let streamInputAdd = htmlToElement("<button>Add</button>");
-        streamInputAdd.onclick = function () {
-            let username = streamInput.getElementsByTagName("input")[0].value.trim();
-            if(username.length == 0) {
-                return;
-            }
-            let stream = [username];
+        streamInputAdd.addEventListener("click", () => {
+            let inputStream = streamInput.value.trim();
+            this._streamManager.addStream(inputStream, this.getUsername());
 
-            window.streamColumns = genColumns(stream, window.urlParser.getSettings(), window.streamColumns, streamInputAdd.closest(".col"));
-            window.urlParser.addStream(username);
             bannerBot.classList.add("hidden");
-            streamInput.getElementsByTagName("input")[0].value = "";
-        };
+            streamInput.value = "";
+        });
 
-        streamInput.appendChild(streamInputAdd);
-        streamInput.appendChild(streamInputCancel);
-        addStreamIcon.onclick = function() {
-            if (Math.floor(window.innerWidth / window.streamColumns.length) < 350) {
-                alert("There is no more space for more streams\nIf you want to add more, increase the size of the window");
+        streamInputWrapper.append(streamInput, streamInputAdd, streamInputCancel);
+        addStreamIcon.addEventListener("click", () => {
+            if (Math.floor(window.innerWidth / this._streamManager.getStreamList().length) < 350) {
+                alert("There is no more space for additional streams\nIf you want to add more, increase the size of the window");
                 return;
             }
-            bannerBot.classList.remove("hidden");
-        };
 
-        bannerTop.appendChild(channelName);
-        bannerTop.appendChild(updateCols);
+            bannerBot.classList.remove("hidden");
+            bannerBot.querySelector("input").focus();
+        });
+
+        bannerTop.append(channelName, updateCols);
         banner.appendChild(bannerTop);
 
         let controls = htmlToElement("<div class='controls'></div>");
         if (this._doAPICalls) {
-            updateCols.appendChild(viewerCount);
-            controls.appendChild(isLiveIcon);
+            updateCols.append(viewerCount, isLiveIcon);
         }
 
-        controls.appendChild(addStreamIcon);
-        controls.appendChild(removeStreamIcon);
-        controls.appendChild(htmlToElement(`<a href='${this.getChannelURL()}' target='_blank' rel='noopener' class='channelButton noselect'>Open</a>`));
+        let userButton = htmlToElement(`<a href='${this.getChannelURL()}' target='_blank' rel='noopener' class='channelButton noselect'>Open</a>`);
+
+        controls.append(addStreamIcon, removeStreamIcon, userButton);
         updateCols.appendChild(controls);
-        bannerBot.appendChild(streamInput);
+        bannerBot.appendChild(streamInputWrapper);
         banner.appendChild(bannerBot);
 
-
-        if (this._doAPICalls) {
-            this._runAPICalls(viewerCount, isLiveIcon);
-        }
+        // For use with StreamManager._runAPICalls()
+        this._viewerCount = viewerCount;
+        this._isLiveIcon = isLiveIcon;
 
         return banner;
     }
@@ -269,41 +273,23 @@ class Stream {
     }
 
     /**
-     * Sets up API polling which will give statistics like viewer count, isLive, etc.
-     * It's only needed to call this method once per streamer since the method contains
-     * a setTimeout() which calls itself again to update the information periodically
-     * @param {dom} viewers span element which will have it's content updated
-     * @param {dom} liveIndicator div that add/remove live class from
+     * Set banner right rightWrapper information
+     * @param {String} text Text to display next to live icon
+     * @param {String} title Hover text for text and live icon
+     * @param {boolean} isLive Sets live icon to red if stream is live
      */
-    _runAPICalls(viewers, liveIndicator) {
-        console.info(`Polling ${this._username} (${this._platform})...`);
+    updateBannerInfo(text, title, isLive) {
+        if(this._banner == null) { return; }
 
-        let request = new XMLHttpRequest();
-        switch(this._platform) {
-            case "twitch":
-                request.open("GET", "https://api.twitch.tv/helix/streams?user_login=" + this._username, true);
+        this._viewerCount.textContent = text;
+        this._viewerCount.title = title;
+        this._isLiveIcon.title = title;
 
-                request.onload = function () {
-                    let data = JSON.parse(this.response);
-
-                    if(data.data.length == 0) {
-                        viewers.textContent = "Not Live";
-                        liveIndicator.classList.remove("live");
-                    } else {
-                        let stream = data.data[0];
-                        // viewers.textContent = `${stream.viewer_count.toLocaleString()} viewers | ${Stream.get_uptime(stream.started_at)}`;
-                        viewers.textContent = `${stream.viewer_count.toLocaleString()} viewers`;
-                        liveIndicator.classList.add("live");
-                    }
-                };
-                request.setRequestHeader("CLIENT-ID", Stream.TWITCH_CLIENT_ID);
-                request.setRequestHeader("Authorization", `Bearer ${Stream.TWITCH_OAUTH_ID}`);
-                break;
+        if(isLive) {
+            this._isLiveIcon.classList.add("live");
+        } else {
+            this._isLiveIcon.classList.remove("live");
         }
-        request.send();
-
-        // Setup next run interval
-        this._apiTimeout = setTimeout(() => { this._runAPICalls(viewers, liveIndicator); }, 60 * 1000);
     }
 
     /**
@@ -321,13 +307,5 @@ class Stream {
             top.querySelector(".rightWrapper .viewerCount").style.removeProperty("display");
             return false;
         }
-    }
-
-    /**
-     * Stop API polling for said stream.
-     * *You can not restart API Polling again once this is called*
-     */
-    stopAPICalls() {
-        clearTimeout(this._apiTimeout);
     }
 }