+++ CHANGELOG.md
@@ -7,3 +7,10 @@ There were a lot of updates prior. Should probably add this to the changelog
 ## 7/25/2020
 
 - Remove Mixer Integration
+
+## 8/16/2020
+
+- Rewrote JS code to be more manageable
+- Fixed adding streams via banner
+- Update Twitch OAuth key
+- Use only a single API Call for stream info rather than n-many